pyproject.toml
[project]
name = "ilyuhin-bot"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[all]~=0.115.12",
    "alembic~=1.15.2",
    "pydantic~=2.11.3",
    "sqlalchemy~=2.0.40",
    "asyncpg~=0.31.0",
    "python-dotenv~=1.1.0",
    "aiogram~=3.22.0",
    "pyngrok"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
